requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# LangChain Ecosystem
langchain>=0.0.280
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Vector Store & Embeddings
faiss-cpu>=1.7.4

# OpenAI
openai>=1.40.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pypdf>=3.17.0

# Text Processing
tiktoken>=0.5.0
regex>=2023.10.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
httpx>=0.24.0
tqdm>=4.66.0
PyYAML>=6.0.1
